Feature/result graph saving

See merge request android/nettest!228

+++ control_client/src/main/java/at/rmbt/client/control/Responses.kt
@@ -147,7 +147,7 @@ data class PingGraphItemResponse(
      * Ping value in milliseconds
      */
     @SerializedName("ping_ms")
-    val durationMillis: Long,
+    val durationMillis: Double,
 
     /**
      * Relative time in milliseconds form the start of the test
